+++ tests/Generator/Fixtures/DefinitionsEnum/Output/Size.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ns\DefinitionsEnum;
+
+enum Size: string {
+    case SMALL = 'small';
+    case BIG = 'big';
+}